Fix signed/unsigned comparison.
svn:sync-xref-src-repo-rev: r98981

+++ src/VBox/Devices/Input/PS2M.cpp
@@ -1,4 +1,4 @@
-/* $Id: PS2M.cpp 54787 2015-03-16 15:06:25Z michal.necasek@oracle.com $ */
+/* $Id: PS2M.cpp 54788 2015-03-16 15:20:58Z noreply@oracle.com $ */
 /** @file
  * PS2M - PS/2 auxiliary device (mouse) emulation.
  */
@@ -562,7 +562,7 @@ static void ps2mReportAccumulatedEvents(PPS2M pThis, GeneriQ *pQueue, bool fAccu
 bool ps2mIsRateSupported(uint8_t rate)
 {
     static uint8_t  aValidRates[] = { 10, 20, 40, 60, 80, 100, 200 };
-    int             i;
+    size_t          i;
     bool            fValid = false;
 
     for (i = 0; i < RT_ELEMENTS(aValidRates); ++i)
